Database for the system of paying for parking
The system of paying for parking. We have a shopping center. There is parking. We can enter to parking by special card, which has a date for us. With this card, we can register our car in the parking. The card includes card id and user id. By user-id, we can get access to the data about the user’s auto.  In registration, we should register our card id, entry and exit dates, parking id. By substruction of entry and exit dates, we can find the total time of parking. And respectively for this time, we should choose tariff. We know the total time, we know tariff, it means, we can find the total amount of due. Also by this database, we can easily find a number of free parking spaces. For this, we just need to count places, where the exit date is null. It means these cars are still there. And we should find the difference between the total number of places and places which are not free